fio.h: Revert type of url from off_t to int.
2001-07-10  Toon Moene  <toon@moene.indiv.nluug.nl>

	* libI77/fio.h: Revert type of url from off_t to int.
	* libI77/dfe.c (c_dfe): Cast offset expression in FSEEK to off_t.
	* libI77/due.c (c_due): Ditto.  (e_rdue): Ditto.
	* libI77/ftell_.c (G77_ftell_0): Cast result of FTELL to integer.
	(G77_fseek_0): Cast offset argument of FSEEK to off_t.

From-SVN: r43910

+++ libf2c/libI77/fio.h
@@ -45,7 +45,7 @@ typedef struct
 	long uinode;
 	int udev;
 #endif
-	off_t url;	/*0=sequential*/
+	int url;	/*0=sequential*/
 	flag useek;	/*true=can backspace, use dir, ...*/
 	flag ufmt;
 	flag urw;	/* (1 for can read) | (2 for can write) */
